Factor out prototyping.tsx.
